Move modules file to Database impl
